Add LocationClient implementation
Created a new LocationClient.cs file to encapsulate the functionality related to fetching
